HydroNex_API/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Swagger-UI==4.11.1
python-dotenv==1.0.1
PyJWT==2.8.0
Werkzeug==3.0.1
google-generativeai==0.3.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
termcolor==2.4.0
pymysql==1.1.0